Bug and quality fix
- Made selected nutrients clearer
- Fixed rounding issue with manual adding

+++ calorie-app/screens/FoodNutritionEditManual/FoodNutritionEditManual.tsx
@@ -38,7 +38,7 @@ const FoodNutritionEditManual = () => {
       .map((item) => {
         return {
           label: item.label,
-          quantity: parseInt(values[item.label]),
+          quantity: parseFloat(values[item.label]),
           unit: item.unit,
         };
       });
